requirement.txt
pillow => 9.4.0
bcrypt => 3.2.2